requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
openai==2.26.0
anthropic==0.84.0
chromadb==0.5.23
sentence-transformers==5.2.3
transformers==4.46.3
torch==2.2.2
konlpy==0.6.0
numpy==1.26.4
pillow==12.1.1
scikit-learn==1.8.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.22
httpx==0.28.1
requests==2.32.5
lxml==6.0.2
bcrypt==5.0.0
google-genai==1.66.0
ibm-watsonx-ai